Add html code for Plain JavaScript with objects
Create a collection that keeps a list of books
Create a function to add a new book to the collection, with title and author.
Create a function to remove a book from the collection
Display all books saved in the collection in the top part of the page.
Make sure that when a user clicks the "Add" button:
A new book is added to the collection.
The new book is displayed in the page.
Make sure that when a user clicks the "Remove" button:
The correct book is removed from the collection.
The correct book disappears from the page.
Make sure that data is preserved in the browser's memory by using localStorage.